// src/models/im/add_user.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface AddUserSchema {
  username: string;
  password: string;
  company?: string;
  role?: number;
  status?: 0 | 1;
  verified?: boolean;
  [k: string]: unknown;
}